Move StorageCreateForm from StorageItem
